server/requirements.txt
flask
Elasticsearch==8.17.0
flask_jwt_extended
python-dotenv